backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
alembic==1.14.1
python-dotenv==1.1.0
supabase==2.15.0
pydantic==2.11.3
gunicorn==23.0.0
